feat(worker): complete secure FileManagerWorker implementation with mTLS + resilient networking
**Core Features:**
- ✅ Self-signed cert generation/store (CurrentUser debug | LocalMachine service)
- ✅ mTLS client auth via WebRequestHandler + auto-cert mgmt
- ✅ Resilient API client: SocketException → WARN retry (no crashes)
- ✅ Topshelf Windows service w/ NetworkService + recovery
- ✅ File sync/rollback w/ PathA→PathB operations

**Key Fixes:**
- 🔧 CS0246 WebRequestHandler → SocketsHttpHandler fallback
- 🔧 CS1501 String.Contains(StringComparison) → IndexOf()
- 🔧 CS0019 TopshelfExitCode cast → (int)comparison
- 🔧 Cert store CryptographicException → UserKeySet|PersistKeySet flags
- 🔧 Network resilience → HttpRequestException.IsNetworkFailure()

**Deployment Ready:**

+++ workers/FileManagerWorker/Program.cs
@@ -123,7 +123,7 @@ namespace FileManagerWorker
                     x.UseNLog();
                 });
 
-                if (exitCode == TopshelfExitCode.Ok)
+                if ((int)exitCode == (int)TopshelfExitCode.Ok)
                 {
                     Console.WriteLine("Service installed successfully.");
                     return 0;
